// src/api/generated/model/dailyRevenueResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MiniCRM API
 * API Documentation for MiniCRM Application
 * OpenAPI spec version: 1.0.0
 */
import type { CarSummary } from './carSummary';
import type { DailyRevenueResponseRemunerationModelType } from './dailyRevenueResponseRemunerationModelType';
import type { DriverSummary } from './driverSummary';

export interface DailyRevenueResponse {
  id?: number;
  date?: string;
  remunerationModelType?: DailyRevenueResponseRemunerationModelType;
  tripCount?: number;
  pricePerTrip?: number;
  driver?: DriverSummary;
  car?: CarSummary;
  kilometersDriven?: number;
  kilometersFrom?: number;
  kilometersTo?: number;
  revenue?: number;
  companyRemuneration?: number;
  driverRemuneration?: number;
  drivingStartTime?: string;
  drivingEndTime?: string;
}